refactor(ci): group runtime estimate tooling
Move the runtime history store, schema, and updater into a dedicated
package while preserving their CI behavior.
